Move portgraph iterators to a submodule
